Fix provisioning gear crashing (#163)
