+++ docs/identifier_provisioning/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to this gear are documented in this file.
 
+## 1.0.8
+* Fix gear crashing when PTID length is greater than 10
+
 ## 1.0.5
 * Updates file suffix check to include the csv file with qc passed requests generated by new enrollment qc workflow.
